New random topic assignment in-progress.

+++ ethereum/contracts/ReputationFactory.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.5.10;
 
-import './Reputation.sol';
-import './Topic.sol';
+import "./Reputation.sol";
+import "./Topic.sol";
 
 contract ReputationFactory {
-    uint deployedReputationCount = 0;
+    uint256 deployedReputationCount = 0;
     mapping(address => address) public deployedReputations;
 
     modifier _topicContractsOnly() {
@@ -15,22 +15,38 @@ contract ReputationFactory {
     }
 
     function createReputation() public {
-        require(deployedReputations[msg.sender] == address(0x0), "A reputation contract has already been created for you.");
+        require(
+            deployedReputations[msg.sender] == address(0x0),
+            "A reputation contract has already been created for you."
+        );
 
         Reputation newReputation = new Reputation(address(this), msg.sender);
         deployedReputations[msg.sender] = address(newReputation);
         deployedReputationCount += 1;
     }
 
-    function addReputation(uint addRep, address sender) public _topicContractsOnly {
+    function addReputation(uint256 addRep, address sender)
+        public
+        _topicContractsOnly
+    {
         Reputation userReputation = Reputation(deployedReputations[sender]);
-        require(userReputation.owner() == sender, "Must be the owner of the Reputation Contract in order to use it.");
+        require(
+            userReputation.owner() == sender,
+            "Must be the owner of the Reputation Contract in order to use it."
+        );
         userReputation.addReputation(addRep);
     }
 
-    function spendReputation(address topicAddress, uint removeRep, address sender) public _topicContractsOnly {
+    function spendReputation(
+        address topicAddress,
+        uint256 removeRep,
+        address sender
+    ) public _topicContractsOnly {
         Reputation userReputation = Reputation(deployedReputations[sender]);
-        require(userReputation.owner() == sender, "Must be the owner of the Reputation Contract in order to use it.");
+        require(
+            userReputation.owner() == sender,
+            "Must be the owner of the Reputation Contract in order to use it."
+        );
         userReputation.spendReputation(topicAddress, removeRep);
     }
 }
